Make sure hitting 'Left' snaps to a caption when it should
Occasionally hitting 'Left' would put you in the middle of a caption,
instead of the beginning of it. This was an artifact of putting an
arbitrary 6 second limit on moving in both directions, but since that's
only to make moving through non-voiced video less weird, it
shouldn't be relied on so much here.