Don't rely on the track element's to populate any captions
I was running into issues where if you specify a negative
offset, any of the captions at the beginning of the video
will get squashed to time 0, and then you won't be able
to retime to them correctly. I was also seeing some weirdness
where captions would actually disappear with negative offsets,
since the track element sometimes gets rid of captions you send
it. Best solution is to not rely on the <track> for anything
but auto-selecting the next active caption while playing.