Prevent backwards navigation from replaying captions
When a line just finishes and you hit backwards, it would
play that line again since it's technically done and
"behind" the current video time. This is a bit awkward
so I'm adding a bit of buffer that makes it so you
have to be well out of range of a finished line before
hitting backwards will go to it again.