Increase contrast ratio of main subtitle to 11
I tried using a different monitor with animebook, and I realized that
the main subtitle actually is pretty gray on most screens. While that's
kinda intentional for when the subtitles are on the black bar, it
looks weird (sorta dirty and washed out?) when the subtitles are
over the show, so I'm increasing the brightness. #bcbcbc on black has
a contrast ratio of 11 (lowest it can be is 1, highest is 21), so I
figure this will be pretty safe when over the black bottom bar, while
also looking reasonably white when over the video.

Didn't mean to be too picky about this originally, but it was really bothering me.