Implement auto pause mode
Fixes #2. I spent some time tweaking the exact times animebook
will detect a place to autopause, and I think I've got it as good as
I can get it with a web-based player. There'll be a bit of weirdness
just with how browsers can only detect changes in the video every
.25 seconds or so, but it should be pretty good if you've got the
timing right with your subs.

Hitting 'A' on your keyboard turns this feature on. I'll probably
add a button or something for it later.