Prevent caption text from blurring out after a screenshot
Not sure why, but occasionally when I screencap an image with
varying browser dimensions, the main caption font gets pixelated
and looks really bad. The video also gets pixelated, so it may
just have to do with what copying an image to the canvas does to
visual elements on the page. I might look into it more later,
but increasing the font size of the main caption seems to make
it better for me.