Don't skip the center of the current caption when seeking screenshots
If you watch past the center of a caption, and then try to get
the screenshot by hitting ',' (key for getting the previous screenshot)
you skip past the center of the caption and end up getting the
previous caption (which you probably don't care about). At first I
thought this made sense, but it's actually really annoying since I
don't know whether to hit , or / whenever I need a screenshot until
I look at the caption I've paused on. I'd rather it "snap" to the
nearest center instead, since that way I can basically always hit
',' after watching a line and it'll take me to the middle of the
caption that I just watched.