Merge duplicate captions
There's almost no reason to have 2 duplicate captions on top of each
other, so I'm merging them together if this happens.

This happens a lot with karaoke done by fansubbers, where they put 2
lines at the same spot with different font colors, positions, etc.
but I can't support that feature of .ass in a web player, so merging's
the best option here.

I guess this will also merge the stereotypical thing where 2 characters
say the same line at the same time, but if that shows up as 1 line
it's not really a huge issue.