Detect keys based on character pressed, not where the key is
Just tried to use a gamepad with animebook and realized
that I was using the wrong api to detect keys. It would
demand you press physical key 's' for example, instead
of another key that produces the letter 's'.

Sorry for any dvorak users out there that maybe ran into this.