Potential fix for weird pixelated subtitles in Chrome
Occasionally when I'm using animebook with Chrome (or Webkit I guess)
the subtitles will get this weird pixelated artifact that seems
to show up randomly on certain subtitles. Weirdly enough, it fixes
itself if I hide subtitles using 'v' and then reshow them again.

I think it's probably a webkit bug, but it seems to be fixed
if I recreate the entire div instead of just the text inside it
when I display the subtitles. Hard to test effectively because
it's a rare bug, but I haven't seen it happen at all since this
fix so I figure it works. I'll update more if I see it again.