Restructure app usage of iframe and runtime messaging
For #19 and #22

I'm making the iframe messaging system more robust, so that if
you drop a file onto the page before the iframe's loaded,
it'll actually wait until the iframe's done before sending
it the file. Before it would just give up.

Along with this came with the realization that using chrome.runtime
is not really the best way of communicating with the background
script, since the background script is loaded in every tab of animebook
you have open in your browser. This means any message I send to it
gets sent to multiple tabs (if you have them open). Ideally, I would
only send messages through the iframe postMessage api, and this would
make it a non-issue, since each tab knows the singular iframe it's supposed to
be communicating with.

But, using the postMessage api has it's own problems, primarily being
that I can't autoplay audio in the background script if I use it.
Autoplaying audio in a chrome extension only seems to work if the call
stack goes through a chrome runtime message. I tried working around this
by making an actual background script (one that's global regardless of
tab you're in) with the sole purpose of playing audio, but having that
background script constantly running revealed a bigger issue: Animebook
Anki Export crashes. A lot. I've just never noticed because it would
always auto reload itself, which it can only do if there's no global
background scripts.

Anyway, I don't know how to fix that since it's ffmpeg crashing the
extension, so I'd rather just leave the architecture mostly as it is
with how it auto-reloads on a crash. Meaning, I'm still going to send
chrome.runtime messages to background tabs unnecessarily, since it's
easier than dealing with other problems I have here. It'll just need
to be smart enough to ignore the message when it's coming from the wrong
tab.