Add lang="ja" attribute
This forces the browser to use the system's default Japanese font for any
elements with lang="ja". Helps prevent the browser from using a Chinese
font by accident. Kinda avoided doing something like this since the
original videobook wanted to be language-agnostic, but that's not
really the goal of the project at this point. Apologies if
you're using animebook to learn Chinese (anyone?). You can still
change the font using css if it's needed.