Improve error messages when ffmpeg fails to start
If pthreads aren't enabled in chrome, I want the error message to be
better. Pthreads are supposed to be on by default, but the nightly builds
of chromium have it turned off for some reason (as of April 2021). Not sure
if it's just because it's a nightly build or if it's something else happening
around now because of recent security changes around chromium, but
just in case new versions of chrome in the next month turn it off,
I want the error message to be there.