Ignore error messages from anki-connect guiBrowse actions
If a guiBrowse action fails, the rest of the app can still work, you
just might risk your card not updating if you have it open in the
browser, similar to when you have the "Do not update Anki browser"
option selected.

For #32 however, you don't even risk that by ignoring the error message,
so there's no downsides to ignoring it. Just log the error and move on.