Update layout to not include visibility toggle
The purpose of this button/bar is kinda misplaced, and its benefit
is pretty trivial when you can hide the sidebar completely anyway. It can
also get confusing, since if you click on it and don't know what it
does, your buttons suddenly disappear and it's not obvious where they
went.