Update extension to manifest V3
Work still in progress. The biggest change still needed on my end
is to make sure that Vue.js on the settings page doesn't use unsafe-eval.

This will demand I build the settings page with some node.js/webpack
setup that can compile my vue templates into render functions
beforehand.

I tried doing that for the toast popups, but thought that webpack was
overkill for it, so I rewrote it in pure js.